xfstests: hard code BSIZE to be 512 bytes
On Linux systems with glibc, DEV_BSIZE defines the sector size for the
Linux kernel, which is 512 since the beginning of time --- and which
we can't really change without breaking a huge amount of kernel code.
Some non-glibc C libraries may not define DEV_BSIZE, and it's simpler
to just use a hard-coded value of 512 bytes.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>