common: rework _require_ext4_mkfs_feature
In all of the places where we need check to see if mkfs.ext4 can
support a set of file system features, we also should be checking to
see if the kernel can support those file system features.  So rename
_require_ext4_mkfs_feature to _require_ext4_feature, and actually
format the file system in $SCRATCH.  To avoid running mkfs twice in
most tests, we will teach the tests to assume that
_require_ext4_feature actually leaves $SCRATCH formatted with a file
system with those features.

Also allow ext4/306 to run on systems where mke2fs doesn't support the
"64bit" option.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>