verity: add common functions for testing fs-verity
Add common functions for testing fs-verity, i.e. filesystem-level
integrity/authenticity protection of readonly files.  Currently
fs-verity is only implemented for f2fs, but ext4 and other filesystems
may support this same API in the future, and the same tests will be
runnable on those other filesystems too.  Running the fs-verity tests
will require the 'fsverity' and 'fsveritysetup' utility programs, which
are not part of xfstests and need to be installed separately.

Signed-off-by: Eric Biggers <ebiggers@google.com>