common/verity: add common functions for testing fs-verity
Add common functions for setting up and testing fs-verity, i.e.
filesystem-level integrity/authenticity protection of readonly files.
Currently fs-verity is only implemented on f2fs and ext4, but other
filesystems may support this same API in the future, and the same tests
should be runnable on those other filesystems too.  Running the
fs-verity tests requires the 'fsverity' utility program, which is not
part of xfstests and needs to be installed separately.

Signed-off-by: Eric Biggers <ebiggers@google.com>