xfs/084: fix test program status collection and processing
On a test VM with 1.2GB memory, I noticed that the test will sometimes
fail because resvtest leaks too much memory and gets OOM killed.  It
would be useful to _notrun the test when this happens so that it doesn't
appear as an intermittent regression.

The exit code processing in this test is incorrect, since "$?" will get
us the exit status of _filter_resv, not $here/src/resvtest.  Fix that
as part of learning to detect a SIGKILL and skip the test.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>