xfs/107: remove long-broken test
This poorly implemented test runs fsstress and embeds xfs_quota output
in the golden output.  This means that it breaks as soon as anyone adds
a new operation to fsstress, since that perturbs the sequence of
operations that it runs, which will make the project quota report output
change.

Normally I'd just fix the test, but the golden output also encodes
output strings that xfs_quota hasn't printed since 2010.  Clearly
nobody's running this test (including me, who has had it turned off for
five+ years) so just get rid of it.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>