misc: refactor getting and setting block device properties via sysfs
Create some helper functions to find the sysfs object corresponding
to a block device, and then add more helpers to get or set properties of
a given block device.  This cleans up the code somewhat and fixes a
problem in generic/108 where it can't handle SCRATCH_DEV pointing to a
partition.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>