common/populate: fix btree-format xattr creation on xfs
Currently, we set a large number of extended attributes when trying to
force the attr fork to be in BTREE format.  This doesn't work reliably
because userspace has no control over where xattr leaf and dabtree
blocks are mapped, and contiguous mappings can prevent the file from
having a btree format attr fork.

However, we /do/ have one small knob for controlling attr fork mappings
in the form of creating remote value xattrs and then deleting them to
leave holes in the mappings.  Create a separate helper function that
exploits this property to try to create a sparse attr fork with enough
mappings to give us the btree attr fork that we want.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>