populate: create fewer holes in directories and xattrs
The same study of aging filesystem metadumps also showed that the rate
of file and xattr deletion is a lot lower than what this script does.
Decrease the deletion interval from every other file/name to every 19th
item to get us to a more reasonable 5% deletion rate, but with a prime
number. ;)

Signed-off-by: Darrick J. Wong <djwong@kernel.org>