generic/711,xfs/537: actually fork these tests for exchange-range
Fork these tests to check the same things with exchange-range as they do
for swapext, since the code porting swapext to commit-range has been
dropped.  generic/711 is forked to generic/1221, and xfs/537 is forked
to xfs/1215.

I was going to fork xfs/789 as well, but it turns out that generic/714
covers this sufficiently so for that one, we just strike fiexchange from
the group tag.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>
Reviewed-by: Christoph Hellwig <hch@lst.de>